requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
WTForms==3.0.1
